+++ packages/twenty-server/src/engine/workspace-manager/workspace-migration-v2/workspace-migration-builder-v2/services/workspace-migration-builder-v2.service.ts
@@ -5,6 +5,7 @@ import { type FromTo } from 'twenty-shared/types';
 import { type FlatObjectMetadataMaps } from 'src/engine/metadata-modules/flat-object-metadata-maps/types/flat-object-metadata-maps.type';
 import { fromFlatObjectMetadataMapsToFlatObjectMetadatas } from 'src/engine/metadata-modules/flat-object-metadata/utils/from-flat-object-metadata-maps-to-flat-object-metadatas.util';
 import { deletedCreatedUpdatedMatrixDispatcher } from 'src/engine/workspace-manager/workspace-migration-v2/utils/deleted-created-updated-matrix-dispatcher.util';
+import { WorkspaceMigrationV2FieldActionsBuilderService } from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-builder-v2/services/workspace-migration-v2-field-actions-builder';
 import { WorkspaceMigrationV2ObjectActionsBuilderService } from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-builder-v2/services/workspace-migration-v2-object-actions-builder.service';
 import { FailedMetadataValidateAndBuild } from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-builder-v2/types/failed-metadata-validate-and-build.type';
 import { WorkspaceMigrationV2 } from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-builder-v2/types/workspace-migration-v2';
@@ -12,7 +13,6 @@ import { computeUpdatedObjectMetadataDeletedCreatedUpdatedFieldMatrix } from 'sr
 import { computeUpdatedObjectMetadataDeletedCreatedUpdatedIndexMatrix } from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-builder-v2/utils/compute-updated-object-metadata-deleted-created-updated-index-matrix.util';
 import { getWorkspaceMigrationV2FieldDeleteAction } from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-builder-v2/utils/get-workspace-migration-v2-field-actions';
 import { getWorkspaceMigrationV2CreateIndexAction } from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-builder-v2/utils/get-workspace-migration-v2-index-actions';
-import { buildWorkspaceMigrationV2FieldActions } from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-builder-v2/workspace-migration-v2-field-actions-builder';
 import { buildWorkspaceMigrationIndexActions } from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-builder-v2/workspace-migration-v2-index-actions-builder';
 
 export type WorkspaceMigrationV2BuilderOptions = {
@@ -27,7 +27,8 @@ export type WorkspaceMigrationBuildArgs = {
 @Injectable()
 export class WorkspaceMigrationBuilderV2Service {
   constructor(
-    private readonly workspaceMigrationV2ObjectActionsBuilder: WorkspaceMigrationV2ObjectActionsBuilderService,
+    private readonly workspaceMigrationV2ObjectActionsBuilderService: WorkspaceMigrationV2ObjectActionsBuilderService,
+    private readonly workspaceMigrationV2FieldActionsBuilderService: WorkspaceMigrationV2FieldActionsBuilderService,
   ) {}
 
   public async validateAndBuild({
@@ -72,7 +73,7 @@ export class WorkspaceMigrationBuilderV2Service {
 
     // Validate and build
     const objectWorkspaceMigrationActions =
-      await this.workspaceMigrationV2ObjectActionsBuilder.validateAndBuildObjectActions(
+      await this.workspaceMigrationV2ObjectActionsBuilderService.validateAndBuildObjectActions(
         {
           fromFlatObjectMetadataMaps,
           createdFlatObjectMetadata,
@@ -84,7 +85,7 @@ export class WorkspaceMigrationBuilderV2Service {
       );
 
     const fieldWorkspaceMigrationActions =
-      buildWorkspaceMigrationV2FieldActions({
+      this.workspaceMigrationV2FieldActionsBuilderService.build({
         inferDeletionFromMissingObjectFieldIndex:
           buildOptions.inferDeletionFromMissingObjectFieldIndex,
         objectMetadataDeletedCreatedUpdatedFields,
